requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.5.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pendulum==2.1.2
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pytzdata==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
win32-setctime==1.0.3
